messenger/requirements.txt
amqp==2.6.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==4.4.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
flower==0.9.4
humanize==0.5.1
idna==3.4
jwcrypto==1.4.2
kombu==4.6.11
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
psycopg2==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.6
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
urllib3==1.26.13
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.1
